ofono2mm: rework bearer management and cleanup the code
initially bearer management was done by erik and it worked "somewhat". then i added some stuff on top of that to cover some edge cases and that worked fine. then someone added try catch blocks around the part that calls doCreateBearer but that made it so that all errors were silenced. initially this was fine because we had no logging and no one encountered any issues in that part of the code.
then we added logging and that had a syntax error in it so that would fail but because there was a try catch without any logging, we would never know when this was failing. this went unnoticed because the part that had a try catch around it was very rarely called only under certain circumstances
this commit basically decouples everything and cleans it up.

Signed-off-by: Bardia Moshiri <fakeshell@bardia.tech>